refactor: move package of colors inside package of theme and rename package in each file
