+++ designSystem/src/main/java/com/example/designsystem/theme/textStyle/DefaultTextStyle.kt
@@ -1,4 +1,4 @@
-package com.amsterdam.aflami.theme.textStyle
+package com.example.designsystem.theme.textStyle
 
 import androidx.compose.runtime.staticCompositionLocalOf
 import androidx.compose.ui.text.TextStyle
